Move animation from ProfileColor to ProfileColorExpandableArea
